+++ changelog.md
@@ -4,6 +4,10 @@ Note, only relatively significant changes to user-visible functionality will be 
 
 May not be updated consistently.
 
+## 20250811
+
+* Added `ButcherTableauSampler` and `LetMeHelpYouSampler` nodes.
+
 ## 20250726
 
 * Internal refactoring, this shouldn't be user-visible but let me know if something breaks.